Separate supabase client/server so it works with auth properly
