+++ src/app/auth.tsx
@@ -3,7 +3,7 @@
 import React, { useEffect } from "react";
 import { usePathname, useRouter } from "next/navigation";
 import { createClientComponentClient } from "@supabase/auth-helpers-nextjs";
-import { DatabaseAutoType } from "@/supabase";
+import { DatabaseAutoType } from "@/database_utils";
 
 export const AuthRouting: React.FC<{ children: React.ReactNode }> = ({ children = <></> }) => {
     const supabase = createClientComponentClient<DatabaseAutoType>();
