+++ src/app/lists/CommentBox.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import FreeFormTextInput from "@/components/DataInput/FreeFormTextInput";
-import supabase from "@/supabase";
+import supabase from "@/supabaseClient";
 import Button from "@mui/material/Button";
 import React, { useState } from "react";
 import styled from "styled-components";
