+++ src/app/lists/EditModal.tsx
@@ -5,7 +5,8 @@ import React, { useState } from "react";
 import styled from "styled-components";
 import { SnackBarDiv } from "@/app/lists/ListDataview";
 import TextInput from "@/components/DataInput/FreeFormTextInput";
-import supabase, { Schema } from "@/supabase";
+import supabase from "@/supabaseClient";
+import { Schema } from "@/database_utils";
 import Snackbar from "@mui/material/Snackbar/Snackbar";
 import Alert from "@mui/material/Alert/Alert";
 import Button from "@mui/material/Button/Button";
