+++ src/app/lists/ListDataview.tsx
@@ -9,7 +9,8 @@ import Table, {
 import React, { useState } from "react";
 import styled from "styled-components";
 import EditModal, { EditModalState, listQuantityNoteAndLabels } from "@/app/lists/EditModal";
-import supabase, { Schema } from "@/supabase";
+import supabase from "@/supabaseClient";
+import { Schema } from "@/database_utils";
 import ConfirmDialog from "@/components/Modal/Confirm";
 import Snackbar from "@mui/material/Snackbar/Snackbar";
 import Alert from "@mui/material/Alert/Alert";
