+++ src/components/LoginPanel.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { DatabaseAutoType } from "@/supabase";
+import { DatabaseAutoType } from "@/database_utils";
 import { createClientComponentClient } from "@supabase/auth-helpers-nextjs";
 import { Auth } from "@supabase/auth-ui-react";
 import { ThemeSupa } from "@supabase/auth-ui-shared";
