+++ src/database_utils.ts
@@ -1,15 +1,4 @@
 import { Database } from "@/database_types_file";
-import { createClient } from "@supabase/supabase-js";
-
-const supabase = createClient<Database>(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-        auth: {
-            persistSession: false,
-        },
-    }
-);
 
 export type DatabaseAutoType = Database;
 
@@ -26,5 +15,3 @@ export type InsertSchema = {
 export type UpdateSchema = {
     [key in keyof Tables]: Tables[key]["Update"];
 };
-
-export default supabase;
