+++ src/supabaseClient.ts
@@ -1,0 +1,6 @@
+import { Database } from "@/database_types_file";
+import { createClientComponentClient } from "@supabase/auth-helpers-nextjs";
+
+const supabase = createClientComponentClient<Database>();
+
+export default supabase;
