+++ src/supabaseServer.ts
@@ -1,0 +1,7 @@
+import { Database } from "@/database_types_file";
+import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
+import { cookies } from "next/headers";
+
+const supabase = createServerComponentClient<Database>({ cookies });
+
+export default supabase;
